basic functionality for workitem workflow
