+++ app/imports/api/problems/index.js
@@ -1,3 +1,4 @@
+import './server/ProblemWorkflow.js';
 import './problems-helpers.js';
 import './problems-service.js';
 import './ProblemWorkflow.js';
