+++ app/imports/api/work-items/WorkItemWorkflow.js
@@ -1,0 +1,6 @@
+if (Meteor.isServer) {
+  import WorkItemWorkflow from './server/WorkItemWorkflow.js';
+  export default WorkItemWorkflow;
+} else {
+  export default class WorkItemWorkflow { }
+}
