+++ app/imports/api/work-items/index.js
@@ -1,5 +1,7 @@
 import './server/publications.js';
+import './server/WorkItemWorkflow.js';
 import './methods.js';
 import './work-item-schema.js';
 import './work-item-service.js';
 import './work-items.js';
+import './WorkItemWorkflow.js';
